SHS-024-ADMIN&USER service are done now -gateway is also done with MDC logging and same with others

+++ userService/src/main/java/com/project/userService/Model/Bookings.java
@@ -13,6 +13,6 @@ import java.util.Map;
 @AllArgsConstructor
 public class Bookings {
     private int maxCount;
-    private int rate;
+    private double rate;
     private Map<LocalDate, BookingList> bookingListMap;
 }
